+++ packages/chrome-devtools/modern.config.ts
@@ -1,4 +1,8 @@
+import path from 'path';
 import { appTools, defineConfig } from '@modern-js/app-tools';
+
+const reactPath = path.dirname(require.resolve('react/package.json'));
+const reactDomPath = path.dirname(require.resolve('react-dom/package.json'));
 
 // https://modernjs.dev/en/configure/app/usage
 export default defineConfig({
@@ -19,6 +23,12 @@ export default defineConfig({
       strategy: 'all-in-one',
     },
   },
+  resolve: {
+    alias: {
+      react: reactPath,
+      'react-dom': reactDomPath,
+    },
+  },
   tools: {
     webpack: (config: Record<string, any>) => {
       config.entry = config.entry || {};
@@ -36,5 +46,5 @@ export default defineConfig({
       return config;
     },
   },
-  plugins: [appTools()],
+  plugins: [appTools({ bundler: 'webpack' })],
 });