+++ packages/enhanced/src/rspack.ts
@@ -1,6 +1,7 @@
 export {
   ModuleFederationPlugin,
   GetPublicPathPlugin,
+  TreeShakingSharedPlugin,
   PLUGIN_NAME,
 } from '@module-federation/rspack/plugin';
 export { createModuleFederationConfig } from '@module-federation/sdk';