+++ packages/modernjs/src/ssr-runtime/SSRLiveReload.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 export function SSRLiveReload() {
   if (process.env.NODE_ENV !== 'development') {
     return null;